Move register.js to test dir where it makes more sense
