+++ test/register.js
